# processor/internal/gateway/docs/swagger.yaml
basePath: /
definitions:
  handler.APIError:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  pb.Account:
    properties:
      account_id:
        description: Unique identifier for the trading account.
        type: string
      account_type:
        description: Type of account (e.g., Stock, Future).
        type: string
      broker_id:
        description: Identifier for the broker.
        type: string
      person_id:
        description: Unique identifier for the account holder.
        type: string
      signed:
        description: Whether the account has signed the trading agreement.
        type: boolean
      username:
        description: User's name associated with the account.
        type: string
    type: object
...
  pb.Action:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-comments:
      Action_ACTION_BUY: Buy. 買進
      Action_ACTION_SELL: Sell. 賣出
    x-enum-descriptions:
    - ""
    - Buy. 買進
    - Sell. 賣出
    x-enum-varnames:
    - Action_ACTION_UNSPECIFIED
    - Action_ACTION_BUY
    - Action_ACTION_SELL
...
  pb.ComboBase:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/pb.Action'
        description: Buy or Sell action for this leg.
      category:
        type: string
      code:
        type: string
      currency:
        $ref: '#/definitions/pb.Currency'
      day_trade:
        $ref: '#/definitions/pb.DayTrade'
      delivery_date:
        type: string
      delivery_month:
        type: string
      exchange:
        $ref: '#/definitions/pb.Exchange'
      limit_down:
        type: number
      limit_up:
        type: number
      margin_trading_balance:
        type: integer
      multiplier:
        type: integer
      name:
        type: string
      option_right:
        $ref: '#/definitions/pb.OptionRight'
      reference:
        type: number
      security_type:
        $ref: '#/definitions/pb.SecurityType'
      short_selling_balance:
        type: integer
      strike_price:
        type: number
      symbol:
        type: string
      target_code:
        type: string
      underlying_code:
        type: string
      underlying_kind:
        type: string
      unit:
        type: number
      update_date:
        type: string
    type: object
  pb.ComboContract:
    properties:
      legs:
        description: Individual components of the combo contract.
        items:
          $ref: '#/definitions/pb.ComboBase'
        type: array
    type: object
  pb.ComboOrder:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
      action:
        $ref: '#/definitions/pb.Action'
      ca:
        type: string
      custom_field:
        type: string
      id:
        type: string
      octype:
        $ref: '#/definitions/pb.FuturesOCType'
      order_type:
        $ref: '#/definitions/pb.OrderType'
      ordno:
        type: string
      price:
        type: number
      price_type:
        type: string
      quantity:
        type: integer
      seqno:
        type: string
    type: object
  pb.ComboTrade:
    properties:
      contract:
        $ref: '#/definitions/pb.ComboContract'
      order:
        $ref: '#/definitions/pb.ComboOrder'
      status:
        $ref: '#/definitions/pb.OrderStatus'
    type: object
  pb.Contract:
    properties:
      category:
        description: Sector or category.
        type: string
      code:
        description: Security code (e.g., stock number).
        type: string
      currency:
        allOf:
        - $ref: '#/definitions/pb.Currency'
        description: Currency used for trading.
      day_trade:
        allOf:
        - $ref: '#/definitions/pb.DayTrade'
        description: Day trade eligibility.
      delivery_date:
        description: Specific delivery date.
        type: string
      delivery_month:
        description: Delivery month (for derivatives).
        type: string
      exchange:
        allOf:
        - $ref: '#/definitions/pb.Exchange'
        description: Exchange where the security is traded.
      limit_down:
        description: Floor price for the day.
        type: number
      limit_up:
        description: Ceiling price for the day.
        type: number
      margin_trading_balance:
        description: Margin trading limit/balance.
        type: integer
      multiplier:
        description: Price multiplier.
        type: integer
      name:
        description: Full name of the security.
        type: string
      option_right:
        allOf:
        - $ref: '#/definitions/pb.OptionRight'
        description: Call or Put.
      reference:
        description: Yesterday's closing or reference price.
        type: number
      security_type:
        allOf:
        - $ref: '#/definitions/pb.SecurityType'
        description: Commodity category (Stock, Future, etc.).
      short_selling_balance:
        description: Short selling limit/balance.
        type: integer
      strike_price:
        description: Strike price (for options).
        type: number
      symbol:
        description: Ticker symbol.
        type: string
      target_code:
        description: Target underlying code.
        type: string
      underlying_code:
        description: Code of underlying asset.
        type: string
      underlying_kind:
        description: Type of underlying asset.
        type: string
      unit:
        description: Trading unit.
        type: number
      update_date:
        description: Date of last update.
        type: string
    type: object
...
  pb.Currency:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    - 13
    - 14
    - 15
    - 16
    - 17
    - 18
    - 19
    - 20
    format: int32
    type: integer
    x-enum-comments:
      Currency_CURRENCY_AUD: Australian Dollar. 澳幣
      Currency_CURRENCY_CAD: Canadian Dollar. 加幣
      Currency_CURRENCY_CHF: Swiss Franc. 瑞士法郎
      Currency_CURRENCY_CNY: Chinese Yuan. 人民幣
      Currency_CURRENCY_EUR: Euro. 歐元
      Currency_CURRENCY_GBP: British Pound. 英鎊
      Currency_CURRENCY_HKD: Hong Kong Dollar. 港幣
      Currency_CURRENCY_IDR: Indonesian Rupiah. 印尼盾
      Currency_CURRENCY_JPY: Japanese Yen. 日圓
      Currency_CURRENCY_KRW: South Korean Won. 韓圓
      Currency_CURRENCY_MYR: Malaysian Ringgit. 馬來西亞幣
      Currency_CURRENCY_NZD: New Zealand Dollar. 紐西蘭幣
      Currency_CURRENCY_PHP: Philippine Peso. 菲律賓披索
      Currency_CURRENCY_SEK: Swedish Krona. 瑞典克朗
      Currency_CURRENCY_SGD: Singapore Dollar. 新加坡幣
      Currency_CURRENCY_THB: Thai Baht. 泰銖
      Currency_CURRENCY_TWD: New Taiwan Dollar. 新台幣
      Currency_CURRENCY_USD: US Dollar. 美元
      Currency_CURRENCY_VND: Vietnamese Dong. 越南盾
      Currency_CURRENCY_ZAR: South African Rand. 南非幣
    x-enum-descriptions:
    - ""
    - New Taiwan Dollar. 新台幣
    - US Dollar. 美元
    - Hong Kong Dollar. 港幣
    - British Pound. 英鎊
    - Australian Dollar. 澳幣
    - Canadian Dollar. 加幣
    - Singapore Dollar. 新加坡幣
    - Swiss Franc. 瑞士法郎
    - Japanese Yen. 日圓
    - South African Rand. 南非幣
    - Swedish Krona. 瑞典克朗
    - New Zealand Dollar. 紐西蘭幣
    - Thai Baht. 泰銖
    - Philippine Peso. 菲律賓披索
    - Indonesian Rupiah. 印尼盾
    - Euro. 歐元
    - South Korean Won. 韓圓
    - Vietnamese Dong. 越南盾
    - Malaysian Ringgit. 馬來西亞幣
    - Chinese Yuan. 人民幣
    x-enum-varnames:
    - Currency_CURRENCY_UNSPECIFIED
    - Currency_CURRENCY_TWD
    - Currency_CURRENCY_USD
    - Currency_CURRENCY_HKD
    - Currency_CURRENCY_GBP
    - Currency_CURRENCY_AUD
    - Currency_CURRENCY_CAD
    - Currency_CURRENCY_SGD
    - Currency_CURRENCY_CHF
    - Currency_CURRENCY_JPY
    - Currency_CURRENCY_ZAR
    - Currency_CURRENCY_SEK
    - Currency_CURRENCY_NZD
    - Currency_CURRENCY_THB
    - Currency_CURRENCY_PHP
    - Currency_CURRENCY_IDR
    - Currency_CURRENCY_EUR
    - Currency_CURRENCY_KRW
    - Currency_CURRENCY_VND
    - Currency_CURRENCY_MYR
    - Currency_CURRENCY_CNY
...
  pb.DayTrade:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      DayTrade_DAY_TRADE_NO: No. 否
      DayTrade_DAY_TRADE_ONLYBUY: Buy first only. 先買後賣
      DayTrade_DAY_TRADE_YES: Yes. 可現沖
    x-enum-descriptions:
    - ""
    - Yes. 可現沖
    - Buy first only. 先買後賣
    - No. 否
    x-enum-varnames:
    - DayTrade_DAY_TRADE_UNSPECIFIED
    - DayTrade_DAY_TRADE_YES
    - DayTrade_DAY_TRADE_ONLYBUY
    - DayTrade_DAY_TRADE_NO
  pb.Deal:
    properties:
      price:
        description: Fill price.
        type: number
      quantity:
        description: Fill quantity.
        type: integer
      seq:
        description: Deal sequence number.
        type: string
      ts:
        description: Timestamp of the fill.
        type: number
    type: object
...
  pb.Exchange:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      Exchange_EXCHANGE_OES: Emerging Stock. 興櫃
      Exchange_EXCHANGE_OTC: Taipei Exchange (Over-the-Counter). 上櫃
      Exchange_EXCHANGE_TAIFEX: Taiwan Futures Exchange. 期貨
      Exchange_EXCHANGE_TSE: Taiwan Stock Exchange. 上市
    x-enum-descriptions:
    - ""
    - Taiwan Stock Exchange. 上市
    - Taipei Exchange (Over-the-Counter). 上櫃
    - Emerging Stock. 興櫃
    - Taiwan Futures Exchange. 期貨
    x-enum-varnames:
    - Exchange_EXCHANGE_UNSPECIFIED
    - Exchange_EXCHANGE_TSE
    - Exchange_EXCHANGE_OTC
    - Exchange_EXCHANGE_OES
    - Exchange_EXCHANGE_TAIFEX
...
  pb.FuturesOCType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      FuturesOCType_FUTURES_OCTYPE_AUTO: Automatic detection. 自動
      FuturesOCType_FUTURES_OCTYPE_COVER: Closing position. 平倉
      FuturesOCType_FUTURES_OCTYPE_DAYTRADE: Intraday day trade. 當沖
      FuturesOCType_FUTURES_OCTYPE_NEW: Opening new position. 新倉
    x-enum-descriptions:
    - ""
    - Automatic detection. 自動
    - Opening new position. 新倉
    - Closing position. 平倉
    - Intraday day trade. 當沖
    x-enum-varnames:
    - FuturesOCType_FUTURES_OCTYPE_UNSPECIFIED
    - FuturesOCType_FUTURES_OCTYPE_AUTO
    - FuturesOCType_FUTURES_OCTYPE_NEW
    - FuturesOCType_FUTURES_OCTYPE_COVER
    - FuturesOCType_FUTURES_OCTYPE_DAYTRADE
...
  pb.GetOrderDealRecordsRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
  pb.GetOrderDealRecordsResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/pb.OrderDealRecord'
        type: array
    type: object
...
  pb.ListComboTradesResponse:
    properties:
      combo_trades:
        items:
          $ref: '#/definitions/pb.ComboTrade'
        type: array
    type: object
...
  pb.ListTradesResponse:
    properties:
      trades:
        items:
          $ref: '#/definitions/pb.Trade'
        type: array
    type: object
...
  pb.OptionRight:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      OptionRight_OPTION_RIGHT_CALL: Call option. 買權
      OptionRight_OPTION_RIGHT_NO: None. 無
      OptionRight_OPTION_RIGHT_PUT: Put option. 賣權
    x-enum-descriptions:
    - ""
    - None. 無
    - Call option. 買權
    - Put option. 賣權
    x-enum-varnames:
    - OptionRight_OPTION_RIGHT_UNSPECIFIED
    - OptionRight_OPTION_RIGHT_NO
    - OptionRight_OPTION_RIGHT_CALL
    - OptionRight_OPTION_RIGHT_PUT
  pb.Order:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/pb.Account'
        description: Account placing the order.
      action:
        allOf:
        - $ref: '#/definitions/pb.Action'
        description: Buy or Sell.
      ca:
        description: Certificate associated with the order.
        type: string
      custom_field:
        description: User-defined data.
        type: string
      daytrade_short:
        description: Whether to allow short day trading.
        type: boolean
      id:
        description: Trade ID (internal).
        type: string
      octype:
        allOf:
        - $ref: '#/definitions/pb.FuturesOCType'
        description: New, Cover, DayTrade (Futures).
      order_cond:
        allOf:
        - $ref: '#/definitions/pb.StockOrderCond'
        description: Cash, MarginTrading, ShortSelling (Stock).
      order_lot:
        allOf:
        - $ref: '#/definitions/pb.StockOrderLot'
        description: Common, Fixing, Odd (Stock).
      order_type:
        allOf:
        - $ref: '#/definitions/pb.OrderType'
        description: ROD, IOC, FOK.
      ordno:
        description: Exchange order number.
        type: string
      price:
        description: Order price.
        type: number
      price_type:
        description: Limit, Market, etc.
        type: string
      quantity:
        description: Number of shares or contracts.
        type: integer
      seqno:
        description: Platform sequence number.
        type: string
    type: object
  pb.OrderDealRecord:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/pb.Action'
        description: Buy or Sell.
      code:
        description: Security code.
        type: string
      price:
        description: Execution price.
        type: number
      quantity:
        description: Execution quantity.
        type: integer
      ts:
        description: Timestamp.
        type: string
    type: object
  pb.OrderStatus:
    properties:
      cancel_quantity:
        description: Quantity cancelled.
        type: integer
      deal_quantity:
        description: Quantity already filled.
        type: integer
      deals:
        description: Individual fill details.
        items:
          $ref: '#/definitions/pb.Deal'
        type: array
      id:
        description: Trade ID.
        type: string
      modified_price:
        description: Last price adjustment.
        type: number
      modified_time:
        description: When the order was last modified.
        type: string
      msg:
        description: Message from the exchange (e.g., errors).
        type: string
      order_datetime:
        description: When the order was created.
        type: string
      order_quantity:
        description: Original order quantity.
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/pb.Status'
        description: Current status (Filled, Pending, etc.).
      status_code:
        description: Detailed status code from exchange.
        type: string
      web_id:
        description: Ordering platform code.
        type: string
    type: object
  pb.OrderType:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      OrderType_ORDER_TYPE_FOK: Fill or Kill. 全部成交或取消
      OrderType_ORDER_TYPE_IOC: Immediate or Cancel. 立即成交或取消
      OrderType_ORDER_TYPE_ROD: Rest of Day. 當日有效
    x-enum-descriptions:
    - ""
    - Rest of Day. 當日有效
    - Immediate or Cancel. 立即成交或取消
    - Fill or Kill. 全部成交或取消
    x-enum-varnames:
    - OrderType_ORDER_TYPE_UNSPECIFIED
    - OrderType_ORDER_TYPE_ROD
    - OrderType_ORDER_TYPE_IOC
    - OrderType_ORDER_TYPE_FOK
...
  pb.SecurityType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      SecurityType_SECURITY_TYPE_FUT: Future. 期貨
      SecurityType_SECURITY_TYPE_IND: Index. 指數
      SecurityType_SECURITY_TYPE_OPT: Option. 選擇權
      SecurityType_SECURITY_TYPE_STK: Stock. 股票
    x-enum-descriptions:
    - ""
    - Index. 指數
    - Stock. 股票
    - Future. 期貨
    - Option. 選擇權
    x-enum-varnames:
    - SecurityType_SECURITY_TYPE_UNSPECIFIED
    - SecurityType_SECURITY_TYPE_IND
    - SecurityType_SECURITY_TYPE_STK
    - SecurityType_SECURITY_TYPE_FUT
    - SecurityType_SECURITY_TYPE_OPT
...
  pb.Status:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    format: int32
    type: integer
    x-enum-comments:
      Status_STATUS_CANCELLED: Cancelled. 已刪單
      Status_STATUS_FAILED: Failed. 失敗
      Status_STATUS_FILLED: Fully filled. 完全成交
      Status_STATUS_INACTIVE: Inactive/Failed. 失敗
      Status_STATUS_PARTFILLED: Partially filled. 部分成交
      Status_STATUS_PENDINGSUBMIT: Pending submission. 傳送中
      Status_STATUS_PRESUBMITTED: Pre-submitted. 預約單
      Status_STATUS_SUBMITTED: Submitted. 委託成功
    x-enum-descriptions:
    - ""
    - Cancelled. 已刪單
    - Fully filled. 完全成交
    - Partially filled. 部分成交
    - Inactive/Failed. 失敗
    - Failed. 失敗
    - Pending submission. 傳送中
    - Pre-submitted. 預約單
    - Submitted. 委託成功
    x-enum-varnames:
    - Status_STATUS_UNSPECIFIED
    - Status_STATUS_CANCELLED
    - Status_STATUS_FILLED
    - Status_STATUS_PARTFILLED
    - Status_STATUS_INACTIVE
    - Status_STATUS_FAILED
    - Status_STATUS_PENDINGSUBMIT
    - Status_STATUS_PRESUBMITTED
    - Status_STATUS_SUBMITTED
  pb.StockOrderCond:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    format: int32
    type: integer
    x-enum-comments:
      StockOrderCond_STOCK_ORDER_COND_CASH: Cash settlement. 現股
      StockOrderCond_STOCK_ORDER_COND_EMERGING: Emerging market stock. 興櫃
      StockOrderCond_STOCK_ORDER_COND_MARGINTRADING: Buying on margin. 融資
      StockOrderCond_STOCK_ORDER_COND_NETTING: Day trade netting. 當沖
      StockOrderCond_STOCK_ORDER_COND_SHORTSELLING: Short selling shares. 融券
    x-enum-descriptions:
    - ""
    - Cash settlement. 現股
    - Day trade netting. 當沖
    - Buying on margin. 融資
    - Short selling shares. 融券
    - Emerging market stock. 興櫃
    x-enum-varnames:
    - StockOrderCond_STOCK_ORDER_COND_UNSPECIFIED
    - StockOrderCond_STOCK_ORDER_COND_CASH
    - StockOrderCond_STOCK_ORDER_COND_NETTING
    - StockOrderCond_STOCK_ORDER_COND_MARGINTRADING
    - StockOrderCond_STOCK_ORDER_COND_SHORTSELLING
    - StockOrderCond_STOCK_ORDER_COND_EMERGING
  pb.StockOrderLot:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    format: int32
    type: integer
    x-enum-comments:
      StockOrderLot_STOCK_ORDER_LOT_BLOCKTRADE: Block trade. 鉅額
      StockOrderLot_STOCK_ORDER_LOT_COMMON: Standard lot. 整股
      StockOrderLot_STOCK_ORDER_LOT_FIXING: Fixed price. 定盤
      StockOrderLot_STOCK_ORDER_LOT_INTRADAY_ODD: Intraday odd lot. 盤中零股
      StockOrderLot_STOCK_ORDER_LOT_ODD: After-hours odd lot. 盤後零股
    x-enum-descriptions:
    - ""
    - Standard lot. 整股
    - Block trade. 鉅額
    - Fixed price. 定盤
    - After-hours odd lot. 盤後零股
    - Intraday odd lot. 盤中零股
    x-enum-varnames:
    - StockOrderLot_STOCK_ORDER_LOT_UNSPECIFIED
    - StockOrderLot_STOCK_ORDER_LOT_COMMON
    - StockOrderLot_STOCK_ORDER_LOT_BLOCKTRADE
    - StockOrderLot_STOCK_ORDER_LOT_FIXING
    - StockOrderLot_STOCK_ORDER_LOT_ODD
    - StockOrderLot_STOCK_ORDER_LOT_INTRADAY_ODD
  pb.SubscribeTradeRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
  pb.SubscribeTradeResponse:
    properties:
      success:
        type: boolean
    type: object
...
  pb.Trade:
    properties:
      contract:
        $ref: '#/definitions/pb.Contract'
      order:
        $ref: '#/definitions/pb.Order'
      status:
        $ref: '#/definitions/pb.OrderStatus'
    type: object
...
  pb.UnsubscribeTradeRequest:
    properties:
      account:
        $ref: '#/definitions/pb.Account'
    type: object
  pb.UnsubscribeTradeResponse:
    properties:
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API Gateway for Shioaji Trading Provider.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Phoenix API
  version: "1.0"
paths:
...
  /api/v1/trades:
    get:
      description: List all trades
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListTradesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Trades
      tags:
      - Trade
  /api/v1/trades/combo:
    get:
      description: List all combo trades
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.ListComboTradesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: List Combo Trades
      tags:
      - Trade
  /api/v1/trades/deals:
    get:
      consumes:
      - application/json
      description: Get deal records for orders
      parameters:
      - description: Deal Records Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.GetOrderDealRecordsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.GetOrderDealRecordsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Get Order Deal Records
      tags:
      - Trade
  /api/v1/trades/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe to trade updates
      parameters:
      - description: Subscribe Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.SubscribeTradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.SubscribeTradeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Subscribe Trade
      tags:
      - Trade
  /api/v1/trades/unsubscribe:
    post:
      consumes:
      - application/json
      description: Unsubscribe from trade updates
      parameters:
      - description: Unsubscribe Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pb.UnsubscribeTradeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.UnsubscribeTradeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIError'
      security:
      - Bearer: []
      summary: Unsubscribe Trade
      tags:
      - Trade
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"